FLGO-218 deploy in testing

+++ sources/server/internal/app/memcache.go
@@ -14,6 +14,9 @@ const (
 )
 
 func (s *Services) DeleteUserCache(ctx context.Context, userId uint64) error {
+	if s.memcache == nil {
+		return nil
+	}
 	err := s.memcache.Delete(ctx, fmt.Sprintf(userKey, userId))
 	if err != nil && !memcached.IsNotFoundError(err) {
 		return err
@@ -22,18 +25,21 @@ func (s *Services) DeleteUserCache(ctx context.Context, userId uint64) error {
 }
 
 func (s *Services) GetBlogArticleLikeCountCache(ctx context.Context, articleId uint64) (uint64, error) {
+	if s.memcache == nil {
+		return 0, nil
+	}
 	likeCountBytes, err := s.memcache.Get(ctx, fmt.Sprintf(articleLikeCountKey, articleId))
-
 	if err != nil {
 		return 0, err
 	}
-
 	likeCount, err := strconv.Atoi(string(likeCountBytes))
-
 	return uint64(likeCount), err
 }
 
 func (s *Services) SetBlogArticleLikeCountCache(ctx context.Context, articleId, likeCount uint64) error {
+	if s.memcache == nil {
+		return nil
+	}
 	bytes := []byte(strconv.Itoa(int(likeCount)))
 	return s.memcache.Set(ctx, fmt.Sprintf(articleLikeCountKey, articleId), bytes, time.Now().Add(1*time.Hour))
 }
